Completes the previous commit (forgotten files + version number).
Some files are simple copies of jetbrain files. I did not manage to make it better : the idea main class loader and the plugin class loader are differents, so that there are two different instances of the package com.intellij.refactoring.ui (and the same for memberPullUp), so that the package access fails between my classes and original classes.

I tried to use a dependency between the plugin and the main com.intellij package to force the same classloader (without success) and I also tried to include the original code as a library (without success). It is not urgent, but that should be fixed.